+++ app/public/css/style.css
@@ -631,14 +631,14 @@ main {
   width: 90%;
 }
 
-.top {
+.index {
   display: flex;
   align-items: center;
   justify-content: center;
   min-height: 100vh;
 }
 
-.top .wrapper {
+.index .wrapper {
   position: relative;
   width: 100%;
   height: 700px;
@@ -646,7 +646,7 @@ main {
   background-color: #fbfbfb;
 }
 
-.top .wrapper:after {
+.index .wrapper:after {
   display: block;
   position: absolute;
   content: "";
@@ -657,7 +657,7 @@ main {
   width: 100%;
 }
 
-.top .content {
+.index .content {
   position: relative;
   z-index: 2;
   display: flex;
@@ -668,22 +668,22 @@ main {
   letter-spacing: 0.1em;
 }
 
-.top h1 {
+.index h1 {
   width: 350px;
   margin-bottom: 30px;
 }
 
-.top h2 {
+.index h2 {
   text-align: center;
   font-size: 18px;
   margin: 152px 0 20px;
 }
 
-.top a {
+.index a {
   margin-top: 30px;
 }
 
-.top .image {
+.index .image {
   z-index: 1;
   position: absolute;
   width: 1000px;
