+++ app/src/app.js
@@ -1,172 +1,31 @@
+// express読み込み
 const express = require('express');
-const mysql = require('mysql2');
-const session = require('express-session');
-const bcrypt = require('bcrypt');
 const app = express();
 
+// モジュール読み込み
+const session = require('./routes/session');
+const indexRouter = require('./routes/index');
+const loginRouter = require('./routes/login');
+const logoutRouter = require('./routes/logout');
+const signupRouter = require('./routes/signup');
+const listRouter = require('./routes/list');
+const articleRouter = require('./routes/article');
+
+// 共通の設定
 app.set('views', './src/views')
 app.use(express.static('public'));
 app.use(express.urlencoded({ extended: false }));
 
-const connection = mysql.createConnection({
-  host: process.env.MYSQL_SERVERE,
-  user: process.env.MYSQL_USER,
-  password: process.env.MYSQL_PASSWORD,
-  database: process.env.MYSQL_DATABASE
-});
 
-connection.connect((err) =>{
-	if (err) throw err;
-	console.log('Connected');
-});
+// ルーティング
+app.use(session);
+app.use('/', indexRouter);
+app.use('/list', listRouter);
+app.use('/article', articleRouter);
+app.use('/signup', signupRouter);
+app.use('/login', loginRouter);
+app.use('/logout', logoutRouter);
 
-app.use(
-  session({
-    secret: 'my_secret_key',
-    resave: false,
-    saveUninitialized: false,
-  })
-);
-
-app.use((req, res, next) => {
-  if (req.session.userId === undefined) {
-    res.locals.username = 'ゲスト';
-    res.locals.isLoggedIn = false;
-  } else {
-    res.locals.username = req.session.username;
-    res.locals.isLoggedIn = true;
-  }
-  next();
-});
-
-app.get('/', (req, res) => {
-  res.render('top.ejs');
-});
-
-app.get('/list', (req, res) => {
-  connection.query(
-    'SELECT * FROM articles',
-    (error, results) => {
-      res.render('list.ejs', { articles: results });
-    }
-  );
-});
-
-app.get('/article/:id', (req, res) => {
-  const id = req.params.id;
-  connection.query(
-    'SELECT * FROM articles WHERE id = ?',
-    [id],
-    (error, results) => {
-      res.render('article.ejs', { article: results[0] });
-    }
-  );
-});
-
-app.get('/signup', (req, res) => {
-  res.render('signup.ejs', { errors: [] });
-});
-
-app.post('/signup',
-  (req, res, next) => {
-    console.log('入力値の空チェック');
-    const username = req.body.username;
-    const email = req.body.email;
-    const password = req.body.password;
-    const errors = [];
-
-    if (username === '') {
-      errors.push('ユーザー名が空です');
-    }
-
-    if (email === '') {
-      errors.push('メールアドレスが空です');
-    }
-
-    if (password === '') {
-      errors.push('パスワードが空です');
-    }
-
-    if (errors.length > 0) {
-      res.render('signup.ejs', { errors: errors });
-    } else {
-      next();
-    }
-  },
-  (req, res, next) => {
-    console.log('メールアドレスの重複チェック');
-    const email = req.body.email;
-    const errors = [];
-    connection.query(
-      'SELECT * FROM users WHERE email = ?',
-      [email],
-      (error, results) => {
-        if (results.length > 0) {
-          errors.push('ユーザー登録に失敗しました');
-          res.render('signup.ejs', { errors: errors });
-        } else {
-          next();
-        }
-      }
-    );
-  },
-  (req, res) => {
-    console.log('ユーザー登録');
-    const username = req.body.username;
-    const email = req.body.email;
-    const password = req.body.password;
-    bcrypt.hash(password, 10, (error, hash) => {
-      connection.query(
-        'INSERT INTO users (username, email, password) VALUES (?, ?, ?)',
-        [username, email, hash],
-        (error, results) => {
-          req.session.userId = results.insertId;
-          req.session.username = username;
-          res.redirect('/list');
-        }
-      );
-    });
-  }
-);
-
-app.get('/login', (req, res) => {
-  res.render('login.ejs');
-});
-
-app.post('/login', (req, res) => {
-  const email = req.body.email;
-  connection.query(
-    'SELECT * FROM users WHERE email = ?',
-    [email],
-    (error, results) => {
-      if (results.length > 0) {
-        // 定数plainを定義してください
-        const plain = req.body.password;
-        const hash = results[0].password;
-
-        bcrypt.compare(plain, hash, (error, isEqual) => {
-          // ここから削除
-          if (isEqual) {
-            req.session.userId = results[0].id;
-            req.session.username = results[0].username;
-            res.redirect('/list');
-          } else {
-            res.redirect('/login');
-          }
-        });
-
-      } else {
-        res.redirect('/login');
-      }
-    }
-  );
-});
-
-app.get('/logout', (req, res) => {
-  req.session.destroy(error => {
-    res.redirect('/list');
-  });
-});
-
+// ポート設定
 const port = process.env.PORT || 3000
 app.listen(port, () => console.log(`server started port ${port}`));
