+++ app/src/routes/article.js
@@ -1,0 +1,17 @@
+const express = require('express');
+const connection = require('./db');
+const router = express.Router();
+
+router.get('/:id', (req, res) => {
+  const id = req.params.id;
+  connection.query(
+    'SELECT * FROM articles WHERE id = ?',
+    [id],
+    (error, results) => {
+      res.render('article.ejs', { article: results[0] });
+    }
+  );
+});
+
+module.exports = router;
+
