+++ app/src/routes/logout.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/', (req, res) => {
+  req.session.destroy(error => {
+    res.redirect('/list');
+  });
+});
+
+module.exports = router;
