+++ app/src/routes/session.js
@@ -1,0 +1,24 @@
+const express = require('express');
+const session = require('express-session');
+const router = express.Router();
+
+router.use(
+  session({
+    secret: 'my_secret_key',
+    resave: false,
+    saveUninitialized: false,
+  })
+);
+
+router.use((req, res, next) => {
+  if (req.session.userId === undefined) {
+    res.locals.username = 'ゲスト';
+    res.locals.isLoggedIn = false;
+  } else {
+    res.locals.username = req.session.username;
+    res.locals.isLoggedIn = true;
+  }
+  next();
+});
+
+module.exports = router;
