+++ app/src/views/index.ejs
@@ -6,7 +6,7 @@
     <link rel="stylesheet" href="/css/style.css">
   </head>
   <body>
-    <div class="top">
+    <div class="index">
       <div class="wrapper">
         <div class="content">
           <h2>わんこの学びブログ</h2>
